adblockplustests/chrome/content/tests/test_requestList.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for request data</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="application/x-javascript;version=1.7" src="../httpd.js"></script>
  <script type="text/javascript; version=1.7" src="/content/tests/SimpleTest/specialpowersAPI.js"></script>
  <script type="text/javascript; version=1.7" src="/content/tests/SimpleTest/SpecialPowersObserverAPI.js"></script>
  <script type="text/javascript; version=1.7" src="/content/tests/SimpleTest/ChromePowers.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">
    <iframe id="frame" onload="doRunTest()"></iframe>
  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    let {RequestNotifier} = require("requestNotifier");
    let {Policy} = require("contentPolicy");
    prepareFilterComponents();
    preparePrefs();

    let currentTest = -1;

    let tests = [
      ["about:blank", runBlankChecks],
      ["http://127.0.0.1:1234/test1", runTest1Checks],
      ["http://127.0.0.1:1234/test2", runTest2Checks]
    ];

    function start()
    {
      window.addEventListener("unload", stop, false);
      server.start(1234);

      server.registerPathHandler("/test1", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<img id='image' src='test.png'>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test2", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script id='script' src='test.js'><" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      runNextTest();
    }

    function runNextTest()
    {
      currentTest++;
      if (currentTest >= tests.length)
      {
        SimpleTest.finish();
        return;
      }

      let [url, testFunction] = tests[currentTest];
      $("frame").src = url;
    }

    function doRunTest()
    {
      if (currentTest < 0)
        return;

      let [url, testFunction] = tests[currentTest];

      let wndScanComplete = false;
      let wndData = [];
      let requestNotifier1 = new RequestNotifier(window, function(wnd, node, entry, scanComplete)
      {
        if (scanComplete)
        {
          wndScanComplete = true;
          requestNotifier1.shutdown();
        }
        else if (entry)
          wndData.push({wnd: wnd, node: node, entry: entry});

        checkScanDone();
      });

      let frameScanComplete = false;
      let frameData = [];
      let requestNotifier2 = new RequestNotifier(window.frames[0], function(wnd, node, entry, scanComplete)
      {
        if (scanComplete)
        {
          frameScanComplete = true;
          requestNotifier2.shutdown();
        }
        else if (entry)
          frameData.push({wnd: wnd, node: node, entry: entry});

        checkScanDone();
      });

      function checkScanDone()
      {
        if (wndScanComplete && frameScanComplete)
        {
          // Fold duplicate entries
          for each (let list in [wndData, frameData])
          {
            let keys = {};
            for (let i = 0; i < list.length; i++)
            {
              let key = " " + list[i].entry.location + " " + list[i].entry.type + " " + list[i].entry.docDomain;
              if (key in keys)
              {
                list.splice(keys[key], 1);
                i--;
              }
              keys[key] = i;
            }
          }

          testFunction(wndData, frameData);
          runNextTest();
        }
      }
    }

    function runBlankChecks(wndData, frameData)
    {
      is(wndData.length, 0, "No data associated with the window that didn't load anything");
      is(frameData.length, 0, "No data associated with an empty frame");
      is(RequestNotifier.getDataForNode($("frame"), true), null, "No data associated with an empty frame node");
    }

    function runTest1Checks(wndData, frameData)
    {
      ok(RequestNotifier.getDataForNode($("frame"), true), "Frame node has data associated with it");
      let frameNodeData = RequestNotifier.getDataForNode($("frame"), true)[1];

      ok(RequestNotifier.getDataForNode($("frame").contentDocument.getElementById("image"), true), "Image node inside the frame has data associated with it");
      let imageNodeData = RequestNotifier.getDataForNode($("frame").contentDocument.getElementById("image"), true)[1];

      is(wndData.length, 2, "Number of entries in the window data");
      if (wndData.length > 0)
      {
        is(wndData[0].node, $("frame"), "First entry in the window data should be the frame");
        is(wndData[0].entry.toSource(), frameNodeData.toSource(), "Data received for the frame should be identical to the result of getDataForNode");
      }
      if (wndData.length > 1)
      {
        is(wndData[1].node, $("frame").contentDocument.getElementById("image"), "Second entry in the window data should be the image");
        is(wndData[1].entry.toSource(), imageNodeData.toSource(), "Data received for the image should be identical to the result of getDataForNode");
      }

      is(frameData.length, 1, "Number of entries in the frame data");
      if (frameData.length > 0)
      {
        is(frameData[0].node, $("frame").contentDocument.getElementById("image"), "Only entry in the frame data should be the image");
        is(frameData[0].entry.toSource(), imageNodeData.toSource(), "Data received for the image should be identical to the result of getDataForNode");
      }

      is(frameNodeData.type, Policy.type.SUBDOCUMENT, "Frame node data: type property");
      is(frameNodeData.typeDescr, "SUBDOCUMENT", "Frame node data: typeDescr property");
      is(frameNodeData.docDomain, "mochikit", "Frame node data: docDomain property");
      is(frameNodeData.thirdParty, true, "Frame node data: thirdParty property");
      is(frameNodeData.location, "http://127.0.0.1:1234/test1", "Frame node data: location property");
      is(frameNodeData.filter, null, "Frame node data: filter property");

      is(imageNodeData.type, Policy.type.IMAGE, "Image node data: type property");
      is(imageNodeData.typeDescr, "IMAGE", "Image node data: typeDescr property");
      is(imageNodeData.docDomain, "127.0.0.1", "Image node data: docDomain property");
      is(imageNodeData.thirdParty, false, "Image node data: thirdParty property");
      is(imageNodeData.location, "http://127.0.0.1:1234/test.png", "Image node data: location property");
      is(imageNodeData.filter, null, "Image node data: filter property");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      ok(RequestNotifier.getDataForNode($("frame"), true), "Frame node still has data associated with it after garbage collection");
      ok(RequestNotifier.getDataForNode($("frame").contentDocument.getElementById("image"), true), "Image node still has data associated with it after garbage collection");
    }

    function runTest2Checks(wndData, frameData)
    {
      ok(RequestNotifier.getDataForNode($("frame"), true), "Frame node has data associated with it");
      let frameNodeData = RequestNotifier.getDataForNode($("frame"), true)[1];

      ok(RequestNotifier.getDataForNode($("frame").contentDocument.getElementById("script"), true), "Script node inside the frame has data associated with it");
      let scriptNodeData = RequestNotifier.getDataForNode($("frame").contentDocument.getElementById("script"), true)[1];

      is(wndData.length, 3, "Number of entries in the window data");
      if (wndData.length > 0)
      {
        is(wndData[0].node, $("frame"), "Second entry in the window data should be the frame (new location)");
        is(wndData[0].entry.toSource(), frameNodeData.toSource(), "Data received for the frame should be identical to the result of getDataForNode");
      }
      if (wndData.length > 1)
      {
        is(wndData[1].node, $("frame"), "First entry in the window data should be the frame (old location)");
        isnot(wndData[1].entry.toSource(), frameNodeData.toSource(), "Data received for the frame should not be identical to the result of getDataForNode");
      }
      if (wndData.length > 2)
      {
        is(wndData[2].node, $("frame").contentDocument.getElementById("script"), "Third entry in the window data should be the script");
        is(wndData[2].entry.toSource(), scriptNodeData.toSource(), "Data received for the script should be identical to the result of getDataForNode");
      }

      is(frameData.length, 1, "Number of entries in the frame data");
      if (frameData.length > 0)
      {
        is(frameData[0].node, $("frame").contentDocument.getElementById("script"), "Only entry in the frame data should be the script");
        is(frameData[0].entry.toSource(), scriptNodeData.toSource(), "Data received for the script should be identical to the result of getDataForNode");
      }

      is(frameNodeData.type, Policy.type.SUBDOCUMENT, "Frame node data: type property");
      is(frameNodeData.typeDescr, "SUBDOCUMENT", "Frame node data: typeDescr property");
      is(frameNodeData.docDomain, "mochikit", "Frame node data: docDomain property");
      is(frameNodeData.thirdParty, true, "Frame node data: thirdParty property");
      is(frameNodeData.location, "http://127.0.0.1:1234/test2", "Frame node data: location property");
      is(frameNodeData.filter, null, "Frame node data: filter property");

      is(scriptNodeData.type, Policy.type.SCRIPT, "Script node data: type property");
      is(scriptNodeData.typeDescr, "SCRIPT", "Script node data: typeDescr property");
      is(scriptNodeData.docDomain, "127.0.0.1", "Script node data: docDomain property");
      is(scriptNodeData.thirdParty, false, "Script node data: thirdParty property");
      is(scriptNodeData.location, "http://127.0.0.1:1234/test.js", "Script node data: location property");
      is(scriptNodeData.filter, null, "Script node data: filter property");

      top.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      ok(RequestNotifier.getDataForNode($("frame"), true), "Frame node still has data associated with it after garbage collection");
      ok(RequestNotifier.getDataForNode($("frame").contentDocument.getElementById("script"), true), "Script node still has data associated with it after garbage collection");
    }

    function stop()
    {
      server.stop();
    }

    let server = new nsHttpServer();
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(start);
  </script>
  </pre>
</body>
</html>
